Powered by

Displays information about authors of the shiny app.

powered_by(
  company_name,
  link = "#",
  position = "top right",
  logo = NULL,
  logo_width = "120px",
  logo_height = "auto",
  inline = FALSE
)

Arguments

company_name
character with the creator of the app

link
link to the creator's website

position
character with position of the parameter. Default "top right".

logo
web link to a logo image or name of the image file in ./www

logo_width
width of the logo in pixels

logo_height
height of the logo in pixels

inline
if TRUE, display text and logo on the same line

Value

div with "powered by".